requirements.txt
flask
PyPDF2
Pillow
openai
reportlab